src/consumer/requirements.txt
kafka-python==2.0.2
psycopg2==2.9.3
python-dateutil==2.8.2
SQLAlchemy==1.4.37
databases[postgresql]==0.2.1